requirements.txt
django
djangorestframework
psycopg2-binary
django-phonenumber-field[phonenumbers]
Pillow